refactor: extract runtime tests
